Three changes: 1. disease parameter durations ar ein hours, not days; 2. Use an exponential distribution rather than Poission for disease phases; 3. small Change to testCovid

+++ src/main/java/uk/co/ramp/covid/simulation/CovidParameters.java
@@ -23,10 +23,10 @@ public class CovidParameters {
     }
 
     private static class DiseaseParameters {
-        public Integer meanLatentPeriod = null;
-        public Integer meanAsymptomaticPeriod = null;
-        public Integer meanPhase1DurationMild = null;
-        public Integer meanPhase1DurationSevere = null;
+        public Double meanLatentPeriod = null;
+        public Double meanAsymptomaticPeriod = null;
+        public Double meanPhase1DurationMild = null;
+        public Double meanPhase1DurationSevere = null;
 
         public Double mortalityRate = null;
         public Double childProgressionPhase2 = null;
@@ -63,19 +63,19 @@ public class CovidParameters {
     }
 
     // Getters
-    public int getMeanLatentPeriod () {
+    public double getMeanLatentPeriod () {
         return diseaseParameters.meanLatentPeriod;
     }
 
-    public int getMeanAsymptomaticPeriod () {
+    public double getMeanAsymptomaticPeriod () {
         return diseaseParameters.meanAsymptomaticPeriod;
     }
 
-    public int getMeanPhase1DurationMild () {
+    public double getMeanPhase1DurationMild () {
         return diseaseParameters.meanPhase1DurationMild;
     }
 
-    public int getMeanPhase1DurationSevere () {
+    public double getMeanPhase1DurationSevere () {
         return diseaseParameters.meanPhase1DurationSevere;
     }
